refactoring preparation for doa helper
