+++ src/Main/Base/SettingsWindow.cpp
@@ -1,10 +1,10 @@
 #include <GWCA/Managers/MapMgr.h>
 
-#include <Base/ChatCommands.h>
 #include <Base/HelperBox.h>
 #include <Base/HelperBoxSettings.h>
 #include <Base/MainWindow.h>
 #include <Defines.h>
+#include <Features/General/ChatCommands.h>
 
 #include <imgui.h>
